+++ packages/theme/src/index.ts
@@ -1,6 +1,6 @@
-'use client';
-
 export * from './ThemeProvider';
+export * from './ThemeScript';
 export * from './utils';
+export * from './cookie';
 export type { ThemeProviderProps } from './ThemeProvider';
 export type { TokenPath } from '@gateflow/tokens';